+++ packages/@expo/fingerprint/src/sourcer/Expo.ts
@@ -10,7 +10,7 @@ import semver from 'semver';
 import { resolveExpoAutolinkingCliPath } from '../ExpoResolver';
 import { getExpoConfigLoaderPath } from './ExpoConfigLoader';
 import { SourceSkips } from './SourceSkips';
-import { getFileBasedHashSourceAsync, stringifyJsonSorted } from './Utils';
+import { getFileBasedHashSourceAsync, relativizeJsonPaths, stringifyJsonSorted } from './Utils';
 import type { HashSource, NormalizedOptions } from '../Fingerprint.types';
 import { toPosixPath } from '../utils/Path';
 import { spawnWithIpcAsync } from '../utils/SpawnIPC';
@@ -43,14 +43,8 @@ export async function getExpoConfigSourcesAsync(
     );
     const stdoutJson = JSON.parse(message);
     config = stdoutJson.config;
-    expoConfig = normalizeExpoConfig(config.exp, options);
+    expoConfig = normalizeExpoConfig(config.exp, projectRoot, options);
     loadedModules = stdoutJson.loadedModules;
-    results.push({
-      type: 'contents',
-      id: 'expoConfig',
-      contents: stringifyJsonSorted(expoConfig),
-      reasons: ['expoConfig'],
-    });
   } catch (e: unknown) {
     if (e instanceof Error) {
       console.warn(`Cannot get Expo config from an Expo project - ${e.message}: `, e.stack);
@@ -134,6 +128,14 @@ export async function getExpoConfigSourcesAsync(
   ).filter(Boolean) as HashSource[];
   results.push(...externalFileSources);
 
+  expoConfig = postUpdateExpoConfig(expoConfig, projectRoot);
+  results.push({
+    type: 'contents',
+    id: 'expoConfig',
+    contents: stringifyJsonSorted(expoConfig),
+    reasons: ['expoConfig'],
+  });
+
   // config plugins
   const configPluginModules: HashSource[] = loadedModules.map((modulePath) => ({
     type: 'file',
@@ -145,7 +147,11 @@ export async function getExpoConfigSourcesAsync(
   return results;
 }
 
-function normalizeExpoConfig(config: ExpoConfig, options: NormalizedOptions): ExpoConfig {
+function normalizeExpoConfig(
+  config: ExpoConfig,
+  projectRoot: string,
+  options: NormalizedOptions
+): ExpoConfig {
   // Deep clone by JSON.parse/stringify that assumes the config is serializable.
   const normalizedConfig: ExpoConfig = JSON.parse(JSON.stringify(config));
 
@@ -212,7 +218,22 @@ function normalizeExpoConfig(config: ExpoConfig, options: NormalizedOptions): Ex
     delete normalizedConfig.web?.splash;
   }
 
-  return normalizedConfig;
+  return relativizeJsonPaths(normalizedConfig, projectRoot);
+}
+
+/**
+ * Gives the last chance to modify the ExpoConfig.
+ * For example, we can remove some fields that are already included in the fingerprint.
+ */
+function postUpdateExpoConfig(config: ExpoConfig, projectRoot: string): ExpoConfig {
+  // The config is already a clone, so we can modify it in place for performance.
+
+  // googleServicesFile may contain absolute paths on EAS with file-based secrets.
+  // Given we include googleServicesFile as external files already, we can remove it from the config.
+  delete config.android?.googleServicesFile;
+  delete config.ios?.googleServicesFile;
+
+  return config;
 }
 
 /**